refactor: Update client domain structure to use IAM namespace and improve exception handling
