+++ src/Application/UseCases/Client/DeleteClient.php
@@ -4,8 +4,8 @@ declare(strict_types = 1);
 
 namespace JuniorFontenele\LaravelVaultServer\Application\UseCases\Client;
 
-use JuniorFontenele\LaravelVaultServer\Domains\Client\Exceptions\ClientException;
-use JuniorFontenele\LaravelVaultServer\Domains\Client\Repositories\ClientRepositoryInterface;
+use JuniorFontenele\LaravelVaultServer\Domains\IAM\Client\Contracts\ClientRepositoryInterface;
+use JuniorFontenele\LaravelVaultServer\Domains\IAM\Client\Exceptions\ClientException;
 
 class DeleteClient
 {
