+++ src/Domains/IAM/Client/Client.php
@@ -2,12 +2,11 @@
 
 declare(strict_types = 1);
 
-namespace JuniorFontenele\LaravelVaultServer\Domains\Client\Entities;
+namespace JuniorFontenele\LaravelVaultServer\Domains\IAM\Client;
 
-use JuniorFontenele\LaravelVaultServer\Domains\Client\Exceptions\ClientException;
-use JuniorFontenele\LaravelVaultServer\Domains\Client\ValueObjects\AllowedScopes;
-use JuniorFontenele\LaravelVaultServer\Domains\Client\ValueObjects\ClientId;
-use JuniorFontenele\LaravelVaultServer\Domains\Client\ValueObjects\ProvisionToken;
+use JuniorFontenele\LaravelVaultServer\Domains\IAM\Client\Exceptions\ClientException;
+use JuniorFontenele\LaravelVaultServer\Domains\IAM\Client\ValueObjects\AllowedScopes;
+use JuniorFontenele\LaravelVaultServer\Domains\IAM\Client\ValueObjects\ProvisionToken;
 
 class Client
 {
